Introduced another type, BOOL

+++ counting-language-base/src/main/java/com/github/mhdirkse/countlang/ast/PrintStatement.java
@@ -22,7 +22,7 @@ public final class PrintStatement extends Statement implements CompositeNode {
 
     @Override
     public void execute(final ExecutionContext ctx) {
-        String result = ((Integer) expression.calculate(ctx)).toString();
+        String result = expression.calculate(ctx).toString();
         ctx.output(result);
     }
